+++ src/main.rs
@@ -1,12 +1,24 @@
+mod alignment;
 mod app;
+mod contig;
+mod contig_collection;
+mod cytoband;
 mod error;
+mod feature;
 mod helpers;
-mod models;
+mod message;
+mod reference;
+mod region;
+mod register;
 mod rendering;
 mod repository;
+mod sequence;
 mod settings;
 mod states;
+mod strand;
+mod track;
 mod traits;
+mod window;
 use app::App;
 use clap::Parser;
 use error::TGVError;
