--- src/models/mod.rs
@@ -1,13 +1,0 @@
-pub mod alignment;
-pub mod contig;
-pub mod contig_collection;
-pub mod cytoband;
-pub mod feature;
-pub mod message;
-pub mod reference;
-pub mod region;
-pub mod register;
-pub mod sequence;
-pub mod strand;
-pub mod track;
-pub mod window;
