+++ src/traits.rs
@@ -1,4 +1,4 @@
-use crate::models::contig::Contig;
+use crate::contig::Contig;
 
 pub trait GenomeInterval {
     fn contig(&self) -> &Contig;
